requirements.txt
Flask
requests
dnspython
openpyxl
python-dotenv